fdt: Link's VGA option ROM is important
On some systems, we require the VGA option ROM to be loaded before VbInit()
is called so we can display BIOS screens, but may not want it for normal
mode. Other systems may not care.

This adds the oprom-matters flag to indicate that it's important to Link.

BUG=chrome-os-partner:8789
TEST=manual

Nothing should change with this commit. It just adds a flag that isn't yet
used.

Change-Id: Ie2255a3a5d41c0280f92f9a3a88acf579e43d32a
Signed-off-by: Bill Richardson <wfrichar@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/27126
Reviewed-by: Randall Spangler <rspangler@chromium.org>