cros: Refactor cros_fdtdec_flashmap() to work sequentially
Rather than building node names and repeatly searching the fdt for
them, scan the flashmap part of the fdt sequentially. This is easier
to code and a little faster to run. It will also make it easier to
add more and more features as required in future.

BUG=chrome-os-partner:10720
TEST=manual
run 'vboot_test fmap' and see that the same values are returned as
before.

Change-Id: Ifb7780f4378424f1714d3e79d961a2f59850e272
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/27692