cros: Fix 4MB flashmap block offsets
These are defined as 64-bit, so should be entered as a single value,
not two 32-bit values. The block offset otherwise comes through
incorrectly.

BUG=none
TEST=manual
Build an image with 'cros_bundle_firmware -b daisy -o out'
$ fdtdump out/updated.dtb |grep block-offset
Manually check that the block offsets are now correct (2 cells, with
the first being the value entered)

            block-offset = <0x00000000 0x00004400>;
            block-offset = <0x00000000 0x00084400>;

Change-Id: I4c1b9322f05da2f16624a912f2ee49f5bc32cca6
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/27693