gen: mkbp: Refactor ec_command() to avoid copy
Now that the various transports can support copy-free operations,
extend this to callers of ec_command() as well.

The only interface change here is to use uint8_t **dinp for the response.
If this is NULL, then *dinp will be set to point to the response
parameters (64-bit aligned). If this is non-NULL, the response
parameters will be copied to *dinp.

BUG=chrome-os-partner:11319
TEST='mkbp test' in U-Boot, see that it succeeds

Change-Id: I8a02edac7935440c5c32140a893bdc911b2e84d6
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/27486
Reviewed-by: Randall Spangler <rspangler@chromium.org>