gen: mkbp: Refactor i2c part to avoid copying
At present all implementations copy data from the internal mkbp
driver buffer to the buffer provided by the caller. This is often
unnecessary, and even counterproductive, since the caller would
generally rather not provide a buffer at all.

Also, since all drivers do the same sort of copy it makes sense to
to it up in ec_command() if it is needed.

Refactor the i2c part of mkbp so that it returns the location of
the param data received, and ec_command() can do the copy itself.

Later, we will update ec_command() to support this new interface.

BUG=chrome-os-partner:11319
TEST='mkbp test' in U-Boot, see that is succeeds

Change-Id: I200e8308ebb1be6acd3a7bd7f201498acd195127
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/27483
Reviewed-by: Randall Spangler <rspangler@chromium.org>