Read and clear the keyboard recovery event
This uses the new keyboard recovery EC event to determine when
keyboard recovery is requested.  It is paired with a change to the EC
to stop reporting keyboard recovery as a switch, since that didn't
provide a good mechanism for the AP to tell the EC to stop reporting
the switch value.

BUG=chrome-os-partner:10034
TEST=manual

1. boot using power+refresh+esc.  See insert screen.
2. power button to power off
3. power button to power on.  Boots to Chrome OS (not insert screen).

This test requires https://gerrit.chromium.org/gerrit/27508 to have
been committed on the EC side first.

Signed-off-by: Randall Spangler <rspangler@chromium.org>

Change-Id: I332d3e8203f0b60d08a395eb9dcfb19aadb74a33
Reviewed-on: https://gerrit.chromium.org/gerrit/27509
Tested-by: Randall Spangler <rspangler@chromium.org>
Reviewed-by: Bill Richardson <wfrichar@chromium.org>
Commit-Ready: Randall Spangler <rspangler@chromium.org>