U-boot bss needs to be initialized
It turns out that coreboot u-boot startup code never initializes the
.bss segment (which is expected to be wiped out to all zeros to ensure
that all uninitialized static/global variables have the initial value
of zero).

When .bss is not wiped out, the code behaves unpredictably depending
on the random values assumed by the static/global the variables.

This change adds some code to wipe out .bss segment when starting
coreboot targets (other u-boot x86 targets should have .bss
initialized in the appropriate startup code). Strictly speaking this
is not necessary when the first u-boot stage is starting up, but the
overhead is negligible and it is good to be explicit.

BUG=chrome-os-partner:11237
TEST=manual
  . generate Link firmware image:
    emerge-link chromeos-u-boot chromeos-bootimage
  . modify it to enable two step boot
   cd ~/trunk/src/platform/vboot_reference/scripts/image_signing
   sudo ./resign_firmwarefd.sh /build/link/firmware/image-link.bin \
      /tmp/two_bios.bin \
      ../../tests/devkeys/firmware_data_key.vbprivk \
      ../../tests/devkeys/firmware.keyblock \
      ../../tests/devkeys/dev_firmware_data_key.vbprivk \
      ../../tests/devkeys/firmware.keyblock \
      ../../tests/devkeys/kernel_subkey.vbpubk 1 0
  . program /tmp/two_bios.bin on a Link target and restart it
  . observe the target bring up ChromeOS (it was getting stuck in boot
    before this change)

Change-Id: Id3768f69c783e5a463344666ae0365c805262df0
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/27782
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>