cros: Support x86-style flashmap nodes
Nodes with @ are still used on x86, and although they serve no purpose,
they need to be supported. Also, on x86 we do not need a block offset.

This was broken by the recent cros_fdtdec_flashmap() refactor which
was accidentally commited before testing on link.

BUG=none
TEST=manual:
build and boot on link, see that it makes it to the kernel now

Change-Id: I82e754f26f7df307ab0c0d35f93a6ea1bf5605eb
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/27883
Reviewed-by: Stefan Reinauer <reinauer@google.com>
Tested-by: Stefan Reinauer <reinauer@google.com>
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>