gen: spi flash: Prevent divide by zero error
The 'sf update' command will produce a divide by zero error if the
length of the update is less 100 bytes.  When this is the case, the
'scale' value will be '0', and then the division using 'scale' can
fault.

This initializes the scale to '1', and only changes it when the number
of bytes to be written is at least 200.

BUG=chromium-os:32603
TEST=Updated small amount of simulated SPI flash; coredump banished.
Signed-off-by: Taylor Hutt <thutt@chromium.org>

Change-Id: Iaff97cc4d14e1a95f6865fedf5af13b4df4bfe64
Reviewed-on: https://gerrit.chromium.org/gerrit/27890
Reviewed-by: Simon Glass <sjg@chromium.org>
Commit-Ready: Taylor Hutt <thutt@chromium.org>
Tested-by: Taylor Hutt <thutt@chromium.org>