cros: fdt: Add mkbp host args and new param block for link
The EC now supports a new larger parameter block, as well as host
arguments which include a checksum, byte length, version and flags.

Add these to the fdt so that the driver will be able to find them.

Note: it isn't completely clear to me why we don't just put the host
args into the message. Then we could use common code with I2C/SPI.

BUG=chrome-os-partner:11319
TEST=none, yet

Change-Id: Ib64d6300287ad625b17df3ad145592b81b39511c
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/28086
Reviewed-by: Randall Spangler <rspangler@chromium.org>