gen: mkbp: Detect and support the new EC interface
Detect the version supported by the EC, and when available, use the
new features of the interface. These include a checksum on packets in
both directions, a data size and a command version.

BUG=chrome-os-partner:11319
TEST=manual:
Build and boot on link, see that 'mkbp info' still works both on old
and new EC images.
Run 'mkbp test' on an old EC (which did not support the new command
format) and see that the old version works and the new version fails.
Run 'mkbp test' on a new EC and see that the old version fails and the
new version works.

Change-Id: I0fc3a9866a4337f61d3dd94bc5db269c3f3f39c4
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/28087