cros: Add .offset and .length to RW-A and RW-B fmap_firmware_entry
We'll need to know this so we can write-protect the selected RW firmware
before booting. The information is already in the FDT, we just need to
add it to the fmap struct so that U-Boot can use it.

BUG=chrome-os-partner:11215
TEST=manual

There should be no noticeable effect, but after booting, run

  grep fmap /sys/firmware/log

You should see these lines (plus a bunch of others):

  fmap.c: dump_fmap_firmware_entry: rw-a
  fmap.c: dump_fmap_entry: all                  003a0000:000f0000

  fmap.c: dump_fmap_firmware_entry: rw-b
  fmap.c: dump_fmap_entry: all                  00490000:000f0000

Change-Id: I77e2d4efd9b1f64d591bfd271d759afb5a1e79d7
Signed-off-by: Bill Richardson <wfrichar@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/28058
Reviewed-by: Simon Glass <sjg@chromium.org>