Fix VbExEcGetExpectedRW()
Should take a const **, since vboot is expected not to modify the
contents of the returned pointer.

Fix missing break statements.

Fix improper use of the file object (need to own the struct and pass
in a pointer, otherwise the pointer is uninitialized.

Also fix a typo in VbExEcHashRW(), which caused it to fail checking
the hash type.

BUG=chrome-os-partner:11148
TEST=if it builds, it worked.  Unfortunately there's at least one other bug
     causing it to return invalid offset/size.

This change MUST be submitted at the same time as vboot_reference
https://gerrit.chromium.org/gerrit/28145.

Change-Id: I86125566f5f5721dfc563416e5255b4e63564bd6
Signed-off-by: Randall Spangler <rspangler@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/28146