usb: fallback safely when a configuration descriptor is too large
When a USB configuration descriptor was larger than our USB buffer
(512 bytes), we are skipping the full descriptor reading but then we
were still parsing and using it.
Now in that case, it just skips this device enumeration and displays the
appropriate message to the user, so he can fix the buffer if he wants.

This bug was triggered by some UVC webcams which have very large
configuration descriptors (e.g. a couple of kB) describing all their
supported video encodings.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BUG=chrome-os-partner:11693
TEST=on Link EVT with Namuga camera, check we skip the webcam
enumeration and we are no longer seeing memory corruption triggering a
freeze when booting from USB.

Change-Id: I3c9e94d586b49e85033df095c98e6c691a5d8a31
Reviewed-on: https://gerrit.chromium.org/gerrit/28270
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>
Commit-Ready: Vincent Palatin <vpalatin@chromium.org>
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>