gen: mmc: Correct type of 'ext_csd' variables
Using a type of 'char' for the 512 byte buffer holding the 'Ext CSD'
MMC register will generate code that is at the mercy of the compiler's
signed/unsigned base type for 'char'.

This can end up having deleterious affects on the code which extracts
the sector count to calculate the capacity of the device.

Changing the type to 'unsigned char' ends up treating the data as raw
bytes with no sign extension.

BUG=chromium-os:32603
TEST=With sandbox, mmc size is 14.7GiB rather than 16EiB
Signed-off-by: Taylor Hutt <thutt@chromium.org>

Change-Id: I0988344b473223df6f1c03be802afdc496a37a1e
Reviewed-on: https://gerrit.chromium.org/gerrit/28286
Tested-by: Taylor Hutt <thutt@chromium.org>
Commit-Ready: Taylor Hutt <thutt@chromium.org>
Reviewed-by: Taylor Hutt <thutt@chromium.org>