x86: config: Increase max SCSI ID to 2 for coreboot board
Not all devices only use SATA port 0 and in order to boot
we need to be able to scan additional ports.

For now this is conservatively increased from 1 to 2.

Devices that do not have anything in other ports can configure
their port map appropriately and the port will be skipped.

BUG=chrome-os-partner:11118
TEST=manual

1) boot on Lumpy and Link and observe that only port0 is scanned
2) boot on Stout and ensure that port1 is scanned and used
(verification step 2 was not done by me)

Change-Id: I3e663beb0748e9d97c522fcfe57c81cb5e51d130
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/28494
Reviewed-by: Kimarie Hoot <kimarie.hoot@se-eng.com>
Tested-by: Kimarie Hoot <kimarie.hoot@se-eng.com>
Reviewed-by: Stefan Reinauer <reinauer@chromium.org>