exynos: Use fixed size types and a packed struture for the power register map
The size of the basic types used here happen to be the same size as the
actual registers, but that's not guaranteed and is fragile. This change
replaces those with fixed size types who are guaranteed to be the same size
as the registers they describe. It also marks the structure as packed so the
compiler won't introduce padding and moving things away from where they're
supposed to be.

BUG=chrome-os-partner:10799
TEST=Built and booted on Snow.

Change-Id: I7981ebaa77283897a2b28618d5c951f41f4f4a21
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/28548
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>
Commit-Ready: Gabe Black <gabeblack@chromium.org>
Reviewed-by: Doug Anderson <dianders@chromium.org>
Tested-by: Doug Anderson <dianders@chromium.org>