exynos: Add concept of a "bad wake" GPIO to machine parameters
This GPIO will be used in a future CL to detect when a wakeup is
invalid.  An invalid wakeup happens when the state of a certain GPIO
is high during wakeup.  This is used to detect a malicious user who
might be trying to mess with the "is sleeping" GPIO.  We detect this
in U-Boot upon wakeup and reboot the system if we find this attack.

BUG=chrome-os-partner:11639
TEST=With next CL, fake an attack by doing 'mw.l 0x114001d0 1' in
U-Boot before booting.  Run 'powerd_suspend'.  When you try to wake
up system will shut itself down and you'll need to power on again.

Change-Id: If97ccd23ea71619697e9c8d37e5b4967fec77a95
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/28611