gen: sandbox: Add column to boards.cfg; move files to new directories
Adds another column to the sandbox board configuration.  This
additional column changes the location of many files, and so they are
moved to accommodate the build process.

This was found to be necessary to get the vboot system to integrate
with the Sandbox.  The reason is that that CONFIG_CHROMEOS is
reuquired to build vboot.  Enabling CONFIG_CHROMEOS builds the 'cros'
directory, and that directory relies on 'soc_family' being set.
Without the final column in the boards.cfg file, 'soc_family' will not
be set and the build of the 'cros' directory will fail.

This could also be addressed by special casing things in the u-boot
build process, but it's definitely better to use built-in
infrastructure rather than special cases to implement functionality.

The build change to include vboot are in one of the next changes.

BRANCH=none
BUG=chromium-os:32603
TEST=Code is not used yet
Signed-off-by: Taylor Hutt <thutt@chromium.org>

Change-Id: I810a8b90f419e5d57f68a1a1c73ab33cbab4cb60
Reviewed-on: https://gerrit.chromium.org/gerrit/31344
Commit-Ready: Taylor Hutt <thutt@chromium.org>
Reviewed-by: Taylor Hutt <thutt@chromium.org>
Tested-by: Taylor Hutt <thutt@chromium.org>