exynos: Add "bad wake" GPIO detection into wakeup path.
This GPIO will be used in a future CL to detect when a wakeup is
invalid.  An invalid wakeup happens when the state of a certain GPIO
is high during wakeup.  This is used to detect a malicious user who
might be trying to mess with the "is sleeping" GPIO.  We detect this
in U-Boot upon wakeup and shut the system down if we find this attack.

BUG=chrome-os-partner:11639
TEST=Fake an attack by doing 'mw.l 0x114001d0 1' in U-Boot before
booting.  Run 'powerd_suspend'.  When you try to wake up system will
reboot instead of waking from suspend.

Change-Id: Icd020a956f118f488994731cc30704a77c405bd4
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/28612
Reviewed-by: Randall Spangler <rspangler@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>