exynos: config: Add an estimate of pre-SPL time
In order to produce an accurate estimate of boot time, we need to
include all time since reset. We have no way of knowing exactly how
much time passed prior to SPL (when we now start our timer), so use
an estimate of 100ms, measured on a scope.

BUG=chromium-os:31642
TEST=manual:
build and boot on snow, see that SPL time is listed:

 Timer summary in microseconds:
       Mark    Elapsed  Stage
          0          0  reset
    100,000    100,000  spl_start
    794,372    694,372  board_init_f
...

Change-Id: Iacf7d372da9cff97759e8718c2366d9bea448bc0
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/28802
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>