gen: bootstage: Add time accumulation feature
Sometimes we want to add up the amount of time spent in a particular
activity when it is happening in a number of discrete chunks.

Add bootstage_start() to mark the start of an acitivity and
bootstage_accum() to accumulate the time since the last start. Calling
these function in pairs results in the accumulated time being collected.

BUG=chrome-os-partner:10660
TEST=manual, in later CL
Add bootstage_start/accum() around LCD init and see that it appears
in the report:

Accumulated time:
                56,729  lcd

Change-Id: I36bf17152bab5dd0d57bcef2d961b82a77405a34
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/29377