exynos: Move CPU detection code out of header file
This doesn't really need to be inline. There is only one caller
(in gpio.h). Move it to a C file.

Also there is really no need to have this in the lowlevel init, so
move it to arch_cpu_init(), which we will enable in a later commit.

BUG=chrome-os-partner:12153
TEST=manual
build and boot to kernel on snow

Change-Id: Ib2200a5815b7ba80b4b282bf138bc699afa59118
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/29539
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>