gen: ahci: Increase command completion timeout
With a cold boot in developer mode on a Parrot device the disk
spins down in the 30 seconds that it is sitting at the prompt,
and so subsequent disk acceses need to be able to wait for the
device to spin up again.

5000ms was chosen to be high enough so as not to fail for a
disk that does not have other issues.

BUG=chrome-os-partner:11370
TEST=manual

1) put Parrot device in developer mode
2) power off the device
3) power on the device again and let it sit at the
developer warning screen for the full 30 seconds
4) ensure that the device boots instead of going to recovery

Change-Id: I8c2c03c057db642c6db81a284dc49c3a67b88904
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/28873
Reviewed-by: Walter Murphy <wmurphy@google.com>
Reviewed-by: Marc Jones <marc.jones@se-eng.com>
Tested-by: Marc Jones <marc.jones@se-eng.com>