Don't keep gbb information in the device tree
As it has been discovered recently, keeping kernel load address and
size information in the device tree is harmful, as the kernel
occasionally could be larger than the allotted size, or could require
loading to a different address. But there are other items address and
size for which is hardcoded in the device tree for no apparent reason.

For instance the gbb structure. On x86 platforms it is not even ever
copied into SRAM - it is mapped directly from the SPI flash. But the
x86 device tree includes the gbb space allocation. This change gets
rid of this definition and fixe one code location which uses gbb.

BUG=chrome-os-partner:11994
TEST=manual
 . build and boot on Link the 'legacy image'
   (/build/link/firmware/nv_image-link.bin)
 . reboot the device, and when at the 'boot >' prompt run the
   'vbexport_test display' command
 . observe proper bit patterns from GBB displayed on the screen

Change-Id: I80f98b1dc9495c937de83a6155eec76146ba866a
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/29117