When appropriate get kernel load address/size information from vboot
Presently kernel load address and buffer size are programmed in the
u-boot device tree. There is no reason for this: the address and size
are part of the vboot encapsulation headers. Duplicating this
information hardcoded in the device tree does not bring any benefit
and is in fact harmful, as it is easy to get out of sync and prevents
the ability to load different payloads at different address.

A better way of doing things is to derive kernel load address and size
from the appropriate vboot header. ARM people object to this, as they
want the very same kernel blob operate on devices with DRAM mapped to
different address ranges.

The suggested solution is to exclude the kernel memory section from
the device tree on the platforms where the load address could be
safely taken from the vboot header. In this case u-boot will pass
address of zero to vboot, which will know to derive the address/size
from the appropriate header and return the derived values to u-boot.

Note that the corssystem blob passed around between u-boot stages
(ro/rw) was mapped 8MB above the default kernel loading address. The
proper solution would be not to allocate arbitrary fixed address at
all, it will be done under a different tracking issue. For now we just
move the fixed address higher to allow for max sized kernel blobs
loaded at the default address.

BUG=chrome-os-partner:11994
TEST=manual
  . testing requires the appropriate vboot patch applied and consists
    of running FAFT on Link.

   - build the new Link firmware image, program it on a Link device,
     restart it and observe it boot ChromeOS

   - observe reported kernel address and size:
    $ ssh <ip addr>  grep kernel_buffer /sys/firmware/log
      cmd_vboot_twostop.c: twostop_main_firmware: - kernel_buffer:      : 00100000
      cmd_vboot_twostop.c: twostop_main_firmware: - kernel_buffer_size: : 00359000

   - observe FAFT succeed:
     $ ./run_remote_tests.sh --board=link --remote=172.22.75.232  --servo \
           control.faft

  - modify for twostop and try booting chromeOS using image-link-tw.bin
    $ cd ~/trunk/src/platform/vboot_reference/scripts/image_signing
    $ sudo ./resign_firmwarefd.sh /build/link/firmware/image-link.bin \
        /build/link/firmware/image-link-tw.bin \
        ../../tests/devkeys/firmware_data_key.vbprivk \
        ../../tests/devkeys/firmware.keyblock  \
        ../../tests/devkeys/dev_firmware_data_key.vbprivk  \
        ../../tests/devkeys/firmware.keyblock \
        ../../tests/devkeys/kernel_subkey.vbpubk 1 0

     observe it boot successfully

  - build a Snow firmware image and try booting the device. observe it
    bring up ChromeOS.

Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Change-Id: I2f276886a95f7e70015f0d6756672659e6f9a574
Reviewed-on: https://gerrit.chromium.org/gerrit/29119
Commit-Ready: Bill Richardson <wfrichar@chromium.org>
Reviewed-by: Bill Richardson <wfrichar@chromium.org>
Tested-by: Bill Richardson <wfrichar@chromium.org>