gen: sandbox: Define memory regions & accessor
This adds an enumeration that is used to refer to the different memory
regions in the Sandbox.  It facilitates dumping out the memory map to
the console on startup, and it allows code to obtain the starting
address of any memory block.  This isn't particularly useful in in
U-boot proper since the device tree will indicate the address, but
this is very helpful for the daemon.

BRANCH=none
BUG=chromium-os:32603
TEST=Code is not used yet
Change-Id: I02aac15c57cd3414225921a99b204ef2183773cb
Signed-off-by: Taylor Hutt <thutt@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/32761
Reviewed-by: Simon Glass <sjg@chromium.org>