cros: Pass FDT's ec-slow-update flag to VbInit()
Some ECs are slow to upgrade. We need to identify those systems with an FDT
flag, and pass it along to vboot so that vboot will know to display the
"please wait" BIOS screen if the EC firmware is being updated. This is only
needed for those slow ECs that already have the ec-software-sync flag in the
FDT.

This CL adds the code that passes the flag. It doesn't change any FDTs.

BUG=chrome-os-partner:12256
TEST=none

No user-visible changes yet, so nothing to test.

Change-Id: I1de92309165d942addd990e3730db7f0e4c57ab6
Signed-off-by: Bill Richardson <wfrichar@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/29384
Reviewed-by: Randall Spangler <rspangler@chromium.org>