sandbox: Improve sandbox serial port keyboard interface
Implements the tstc() interface for the serial driver.  Multiplexing
the console between the serial port and a keyboard uses a polling
method of checking if characters are available; this means that the
serial console must be non-blocking when attempting to read
characters.

BUG=chromium-os:32603
TEST=When a keyboard device is added to the Sandbox, it's possible to provide
input from both the device and the serial port.
Signed-off-by: Taylor Hutt <thutt@chromium.org>

Change-Id: Ic8ec6b6706ce1b6735a1f05d112abd1081c85f6a
Reviewed-on: https://gerrit.chromium.org/gerrit/29608
Reviewed-by: Simon Glass <sjg@chromium.org>
Tested-by: Taylor Hutt <thutt@chromium.org>
Commit-Ready: Taylor Hutt <thutt@chromium.org>