exynos: Put SPL marker in its own special place
The SPL marker currently goes just above the IRAM stack. It seems better to
give it an explicit address.

Also, the initial stack used for SPL should not be SDRAM, as this is not available.

So, create a new CONFIG_SPL_MARKER config option for the address, and use the
same stack (in IRAM) for SPL and the U-Boot early init.

BUG=chrome-os-partner:12153
TEST=manual
build and boot on snow with USB and SPI
see that it correctly detects read-only and read-write firmware
see that kernel boots correctly

Change-Id: Id5bfcb7351a431787f3edb3fcb604cc19800eb2e
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/29543