exynos: Refactor low_level code to clarify paths
The current code repeats itself because it needs to do clock init
for a wake from sleep but not a week from idle.

Refactor the code to use a action mask, to avoid the duplication.
None of this init is required outside SPL, so make it conditional on
that (the next commit moves the whole file to SPL-only).

BUG=chrome-os-partner:12153
TEST=manual:
build and boot to kernel on snow
Check that suspend/resume works correctly
TODO: need to test the idle resume

Change-Id: I84d9db524cda4b1f3f14f8660ce7d77a58353c7b
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/29545
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Reviewed-by: Alim Akhtar <alim.akhtar@samsung.com>