cros: clean up warnings for vboot_twostop()
Cleans up warnings produced when building Sandbox U-Boot.  Without
this change, for the Sandbox U-Boot, a warning is emitted that
vboot_twostop() is defined but not used.  Presently, the Sandbox
U-Boot does not implement a R/W firmware, so jumping to the
non-existent R/W version of the firmware is not possible for now.

BRANCH=none
BUG=chromium-os:32603
TEST=Code is not used yet
Change-Id: I83c5d312017b7e28a9a28c2e9d8678da0926a2ec
Signed-off-by: Taylor Hutt <thutt@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/30595
Reviewed-by: Simon Glass <sjg@chromium.org>