cros: connect 'vboot_flag_id' values and string representations
This change combines the definition of the enumeration 'vboot_flag_id'
and the textual representation of these identifiers.

This makes the code less error prone since the defintion of the
enumerate identifiers and the initialized array containing the strings
will always be properly initialized.  It also makes it easier to see
the relationship between the identifier and the string when they are
in the same place.

BUG=chromium-os:32603
TEST=u-boot still builds; no code changes, no data changes.
Signed-off-by: Taylor Hutt <thutt@chromium.org>
Change-Id: I2faa9065f176fdcfda9069632a3ee3ffa0c51d27
Reviewed-on: https://gerrit.chromium.org/gerrit/29814
Reviewed-by: Simon Glass <sjg@chromium.org>