gen: tpm: Probe TPM again when no transmission for a long time
Verified boot library does several rounds of TPM transmissions, like TPM read,
write, and lock. But a long idle time may be between some of them, like
waiting user for a dev screen, waiting user to plug a USB disk in, etc.
The TPM may go to asleep that time and cause the next transmission failed.

This CL is to fix this issue. Before doing a TPM transmission, we check the
duration between the current one and the last one. If it is longer than 2sec,
we do a probling (sending some traffic to the I2C bus) to wake up TPM.

BUG=chrome-os-partner:12076
TEST=manual
 - Patch this CL to ToT firmware and flash it to Snow;
 - Modify the GBB flag to 0;
 - Switch to dev mode;
 - Power on the device and wait until dev screen timeout;
 - Success to boot to Chrome OS login screen.

Change-Id: Ic49031e201d2383b0532353c2189030a154d1802
Signed-off-by: Tom Wai-Hong Tam <waihong@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/29867