Increase GBB size to 956K to match x86.
We're adding more and bigger BIOS bitmaps, so we need more space to hold
them. There's plenty of room on ARM systems, but we'll only increase the
size to match that on the x86 system so we have the same restrictions on
each.

BUG=chrome-os-partner:12253
TEST=manual

No user-visible changes; everything should still work as before.

Change-Id: Id0b7d0a547c2c4e0fcb729568f6445a5c4a66122
Signed-off-by: Bill Richardson <wfrichar@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/30077
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Commit-Ready: David James <davidjames@chromium.org>