exynos: Add a codec enable GPIO to Daisy and Snow.
Both Daisy and Snow have a GPIO which comes from the Exynos and goes to the
enable pin on their audio codecs. This change sets that GPIO in the device
tree so that startup code already added to smdk5250.c will set it
appropriately.

BUG=chrome-os-partner:10708
TEST=Built and booted on Snow and could interact with the codec without having
to manually set the GPIO to 1.

Change-Id: I943f9b37ccebe07a5e26ad996bcf1c19d8e8713f
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/30164
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>
Commit-Ready: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>