sound: Make U-Boot able to play different frequencies and durations
U-Boot was hardwired to play a sine wave at 400 Hz for 1 second. This CL
changes that in a number of ways. First, it makes the frequency and duration
configurable in the underlying sound_play function. Second, because we can't
use a precomputed sine wave table with arbitrary frequencies, the code
generates a square wave instead. The sound is harsher, but the frequency is
configurable. Third, this CL extends the "sound" command so that it optionally
takes a duration in milliseconds and a frequency in hertz, and defaults to the
old values if none are specified.

BUG=chrome-os-partner:11520
BRANCH=snow
TEST=Built and booted on Snow. Played a number of different durations and
frequencies with "sound play" and heard different and seemingly correct
sounds.

Change-Id: I2458086a61e8e04ab094303e7ce013edba306680
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/30185
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Commit-Ready: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>