x86: Add a command to load VGA rom
In non-verified boot, vboot is not initialized and VGA rom is not loaded
when needed. This command can be used in this case to load VGA rom if it
is needed.

BUG=chrome-os-partner:11400
TEST=Build with DEV_MODE gbb flag. Clear NVRAM and reboot. Check display
is off. Run this command and wait for reboot. Check display is on. Run
again when display is already on. Check no error occurs.

Change-Id: If39cb8309fb579c5270f6f126dadf19035a96ed2
Signed-off-by: Vic Yang <victoryang@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/30372
Reviewed-by: Bill Richardson <wfrichar@chromium.org>