gen: mkbp: Deal with partial i2c transfers
When an i2c transfer fails it may return a positive integer. We must
return -1 on error, so update the return code in this case.

BUG=chrome-os-partner:12530
TEST=manual:
build and boot to U-Boot on snow
'mkbp id' on an EC specially set up to return too few bytes
See that it now gives an error

Change-Id: I182245763a045f0cbbfea4f5c4b7eb1ee703d6fa
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/30275
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>