gen: mkbp: Wait for flash erase to finish before returning
Use the new GET_STATUS command to watch for completion of a flash
erase operation, and only return when it is complete. This avoids
future commands failing because the EC is busy.

BUG=chrome-os-partner:12530
TEST=manual:
build and boot to U-Boot on snow
Use 'mkbp erase' command in next commit, and a special EC with a
ccprintf() in flash_command_erase() to test that the command does
not complete until the EC is done erasing.

Change-Id: Icb6a34855362cad214a8427db3bc7a1657e145a8
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/30277
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>