gen: mkbp: Add mkbp_flash_write() to write to flash
Move the code to write to flash into a function, with a block write
function also. This will be used by a new 'mkbp write' command.

BUG=chrome-os-partner:12530
TEST=manual:
build and boot to U-Boot on snow
Use 'mkbp write' command in next commit to test that we can write to
the flash.
NOTE: Software sync's use of mkbp_flash_update_rw() needs to be tested
also.

Change-Id: Iae0e5c3d21c8471603cd598a80e88ca584134981
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/30279