exynos: spi: Support SPI_PREAMBLE mode
Support interfaces with a preamble before each received message.

We handle this when the client has requested a SPI_XFER_END, meaning
that we must close of the transaction. In this case we read until we
see the preamble (or a timeout occurs), skipping all data before and
including the preamble. The client will receive only data bytes after
the preamble.

BUG=chrome-os-partner:10533
BRANCH=none
TEST=manual
With other changes, see that mkbp works correctly over SPI using
'mkbp id' command.

Change-Id: Ibe7eeb61e03965d5d6b0d4154d317a8d30202038
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/30774