exynos: spi: Support a delay after deactivate
For devices that need some time to react after a spi transaction
finishes, add the ability to set a delay.

Implement this as a delay on the first/next transaction to avoid
any delay in the fairly common case where a SPI transaction is
followed by other processing.

The fdt binding is:

  spi-deactivate-delay: Delay to wait after rising edge of SPI
     transaction before allowing a new transaction to start. The
     SPI CS signal must remain high at least for this long between
     transactions.

BUG=chrome-os-partner:10533
BRANCH=none
TEST=manual
With other changes, see that mkbp works correctly over SPI using
'mkbp id' command.

Change-Id: Iee94738b1228e6de71c4d5a478ea04b30527762f
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/30775
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>