gen: mmc commands: Fix incorrect handling of 'read' & 'write' commands
If a malformed 'read' or 'write' command is issued, the Sandbox U-Boot
can crash because the command-handling code does no error checking on
the number of provided arguments.

This change makes the mmc 'erase', 'read' and 'write' commands only
function if the proper number of arguments are supplied.

Also puts the else assignment at the beginning fo the if() statement
to shortens the generated code.  This removes an unnecessary jump from
the generated code.

BRANCH=none
BUG=chromium-os:32603
TEST=Fixes crash with malformed 'mmc read' and 'mmc write' commands
Signed-off-by: Taylor Hutt <thutt@chromium.org>
Change-Id: I02424b1a7f4e83a0860289f36457097a2d92af83
Reviewed-on: https://gerrit.chromium.org/gerrit/31027
Reviewed-by: Simon Glass <sjg@chromium.org>