Do not initialize device tree twice
On coreboot flavor of u-boot the device tree is read from CBFS while
running in the 'ROM based' mode (even though running from DRAM). A
memory area preserved by coreboot is used for storing the volatile
copy of the device tree.

The relocated instance of u-boot again reads the device tree from the
CBFS, and uses a malloc'ed buffer for that. This is not necessary and
even harmful, as the device tree modifications made while running 'ROM
based' get lost.

This change drops the code retrieving the device tree from CBFS for
the second time.

BRANCH=link
BUG=chrome-os-partner:9512
TEST=manual
  . programmed the new firmware image on a Link device
  . verified that gd->fdt_blob contains a pointer to a valid device
    tree when board_early_init_r() starts (extra code, now removed,
    was used for that)
  . booted ChromeOS

Change-Id: I2cf856cf4356f0513707ce39e4750a44a53a9212
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/31191
Reviewed-by: Stefan Reinauer <reinauer@google.com>
Reviewed-by: Simon Glass <sjg@chromium.org>