arm: power_shutdown() must never return, as advertised.
The AP expects power_shutdown() to cause the system to shutdown immediately
so that the function call never returns. If that doesn't happen, then the
AP continues doing other things that it shouldn't be doing, which may
cause problems.

This change makes power_shutdown() never return.

BUG=chrome-os-partner:12412
BRANCH=snow
TEST=none

This doesn't fix anything in particular, but it reduces the scope of
potential problems by just a little bit.

Change-Id: I36b66fa382550c9f9fc33fe718de20c75a610586
Signed-off-by: Bill Richardson <wfrichar@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/31241
Reviewed-by: Randall Spangler <rspangler@chromium.org>